修复第一次渲染时候会执行两次的bug,增加pccs文件找不到提示信息

+++ class/shutdown.php
@@ -2,9 +2,17 @@
 
 // 执行pcss文件的解析工作，并将文件输出
 if(!isset($arg)){
-    $arg='';
+    $arg=array();
 }
 $pcss_file=dirname($currenFile).'/'.basename($currenFile,'.php').'.pcss';
+
+if(!file_exists($pcss_file)){
+    echo "$pcss_file not found!\n";
+    return;
+}
+
 $result=$latte->renderToString($pcss_file, $arg);
+
 $filename='/'.basename($currenFile,'.php').'.css';
+
 file_put_contents($destPath.$filename, $result);
